Initial Commit
The commit contains all of the progress made from the beginning of January until now. A wireframe for the shops, and product filter has been made, and the account dashboard is being finalized. At the back end of the project, the project includes functions for creating an account, a shop, a listing, and many more (2FA, Searching, etc).